Menu e Tabelas

+++ src/main/java/com/salaodebeleza/model/modells/Departamento.java
@@ -3,6 +3,9 @@ package com.salaodebeleza.model.modells;
 import java.util.List;
 
 
+
+
+
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
@@ -11,17 +14,21 @@ import javax.persistence.Id;
 import javax.persistence.OneToMany;
 import javax.persistence.Table;
 
+import com.salaodebeleza.model.modells.Departamento;
+
+
+
 @Entity
 @Table(name="TAB_DEPARTAMENTO")
 public class Departamento {
 	
-	private Long id;
+	private Integer id;
 	private String nome;
 	
 	private List<Usuario> usuario;
 	
 	
-	public Departamento(Long id, String nome) {
+	public Departamento(Integer id, String nome) {
 		this.id = id;
 		this.nome = nome;
 	}	
@@ -34,11 +41,11 @@ public class Departamento {
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
 	@Column(name="id", nullable = false)
-	public Long getId() {
+	public Integer getId() {
 		return id;
 	}
 	
-	public void setId(Long id) {
+	public void setId(Integer id) {
 		this.id = id;
 	}
